refactor : 폴더 정리 완료
