+++ front-end/src/components/Allrecipe/Allrecipe.js
@@ -1,48 +1,8 @@
 import React from "react";
+import FoodCard from "./FoodCard.js";
 import "./style.css";
 
-function StarsComopnent(props) {
-  if (props.star) {
-    return (
-      <div
-        style={{
-          background: "yellow",
-          width: "50px",
-          height: "50px",
-          border: "solid 1px",
-        }}
-      />
-    );
-  } else {
-    return (
-      <div
-        style={{
-          width: "50px",
-          height: "50px",
-          border: "solid 1px",
-        }}
-      />
-    );
-  }
-}
-function FoodCard(props) {
-  const stars = [0, 0, 0, 0, 0];
-  for (let index = 0; index < props.star; index++) {
-    stars[index] += 1;
-  }
-  const starsComopnents = stars.map((star) => <StarsComopnent star={star} />);
-  return (
-    <div className="allrecipe">
-      <img className="foodimge" src={props.imgSrc} alt="" />
-      <div className="allrecipe_content">
-        <strong className="foodname">{props.name}</strong>
-        <p className="star">{starsComopnents}</p>
-      </div>
-    </div>
-  );
-}
-
-function Allrecipe() {
+export default function Allrecipe() {
   const datas = [
     {
       name: "김치찌개",
@@ -67,5 +27,3 @@ function Allrecipe() {
   ));
   return <div>{foodList}</div>;
 }
-
-export default Allrecipe;
